requirements.txt
fastapi
uvicorn
twilio
python-dotenv
openai
langgraph
pydantic
PyPDF2
